drm/i915: move *_crtc_clock_get() to intel_dpll.c
commit 9d6953335284fc37f25bf8488a15ee9444198248 upstream.

Considering what the functions do, intel_dpll.c is a more suitable
location, and lets us make some functions static while at it.

This also means intel_display.c no longer does any DPIO access.

deepin-Intel-SIG: commit 9d6953335284 drm/i915: move *_crtc_clock_get() to intel_dpll.c.
MTL Graphics backporting - 2024Q2

Link: https://patchwork.freedesktop.org/patch/msgid/20231114104534.4180144-1-jani.nikula@intel.com
[ Quanxian Wang: amend commit log ]
